feat(opctl): WR-162 SP 14 Phase A — extract tier-display helpers to renderer-safe module
Move four pure presentation helpers (`getRequiredTier`, `getTierDisplay`,
`T3_COOLDOWN_MS`, `ConfirmationTierDisplay`) from `confirmation.ts` to a
new `tier-display.ts` (zero `node:*` imports). The package barrel re-routes
the four helpers through `./tier-display.js` so renderer chunks reach the
renderer-safe module first; `confirmation.ts` retains `node:crypto` for
`issueSystemProof` / `hashScope` and is not exercised by the renderer
post-extraction.

Internal direct-import callers (`opctl-service.ts`, `tier-display.test.ts`,
`confirmation-system-proof.test.ts`) re-pointed to `./tier-display.js`
(SDS-M2 reconciliation absorbed at CR).

SUPV-SP14-021 closure; closes SP 10 Note 1b renderer-shim durability vs.
`confirmation.ts` split.
